pyproject.toml
[project]
name = "alx-backend-security"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "djangorestframework>=3.16.0",
]
